requirements.txt
appdirs==1.4.3
asgiref==3.2.7
Django==3.0.5
django-cors-headers==3.2.1
django-extensions==2.2.9
django-rest-framework==0.1.0
djangorestframework==3.11.0
jedi==0.16.0
parso==0.6.2
prompt-toolkit==3.0.5
ptpython==3.0.1
Pygments==2.6.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
wcwidth==0.1.9
whitenoise==5.0.1
